+++ src/main/java/globaloutbreak/model/mutation/MutationFactoryImpl.java
@@ -6,7 +6,7 @@ package globaloutbreak.model.mutation;
 public final class MutationFactoryImpl implements MutationFactory {
 
     @Override
-    public Mutation createMutation(final int cost, final String name, final int increase, final TypeMutation type, 
+    public Mutation createMutation(final int cost, final String name, final float increase, final TypeMutation type, 
                                     final String description) { 
         return new MutationImpl(cost, name, increase, type, description);
     }
